microservices/analytics-service/requirements.txt
# ==========================================
# DJANGO CORE
# ==========================================
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
# ==========================================
# DATABASE
# ==========================================
psycopg2-binary==2.9.9
prisma==0.11.0
# ==========================================
# CACHE & QUEUE
# ==========================================
redis==5.0.1
celery==5.3.4

# ==========================================
# AUTHENTICATION & SECURITY
# ==========================================
PyJWT==2.8.0
djangorestframework-simplejwt==5.2.2

# ==========================================
# CONFIGURATION & ENVIRONMENT
# ==========================================
python-decouple==3.8

# ==========================================
# HTTP CLIENTS
# ==========================================
requests==2.31.0

# ==========================================
# PRODUCTION SERVER
# ==========================================
gunicorn==21.2.0

# ==========================================
# DEVELOPMENT & CODE QUALITY
# ==========================================
black==23.11.0
isort==5.13.2
flake8==6.1.0
pylint-django==2.5.5

# ==========================================
# TESTING
# ==========================================
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.4.0

# ==========================================
# DEBUGGING & DEVELOPMENT TOOLS
# ==========================================
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.17.2

# ==========================================
# DOCUMENTATION
# ==========================================
drf-yasg==1.21.7
coreapi==2.3.3

# ==========================================
# ASYNC & PERFORMANCE
# ==========================================
channels==4.0.0
daphne==4.0.0
